pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepexomir"
version = "0.1.0"
description = "DeepExoMir: Exosome miRNA Target Prediction Platform for Aesthetic Medicine"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10,<3.13"
authors = [
    {name = "XunLian Group R&D"},
]

dependencies = [
    "torch>=2.2",
    "transformers>=4.40",
    "multimolecule>=0.1",
    "biopython>=1.83",
    "streamlit>=1.35",
    "pandas>=2.1",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "requests>=2.31",
    "networkx>=3.2",
    "pyvis>=0.3",
    "plotly>=5.18",
    "tensorboard>=2.15",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "ruff>=0.1",
]
vienna = [
    "ViennaRNA>=2.6",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["deepexomir*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py311"
